+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.4.0
 
 * Adding `#.removeNodeAttribute` & `#.removeEdgeAttribute` that was actually missing from earlier versions.
+* Adding graph attributes.
 * `nodeUpdated` & `edgeUpdated` changing to `nodeAttributesUpdated` & `edgeAttributesUpdated` respectively.
 
 ## 0.3.0